+++ lib/fb_news_service.dart
@@ -4,8 +4,9 @@ class FbNewsService {
   Future<http.Response> getFeed({
     String pageId,
     String token,
+    int limit = 20,
   }) async {
     return await http.get(
-        "https://graph.facebook.com/v9.0/$pageId/feed?fields=id,full_picture,created_time,message,attachments,likes{id},shares,permalink_url,from&access_token=$token");
+        "https://graph.facebook.com/v9.0/$pageId/feed?fields=id,full_picture,created_time,message,attachments,likes{id},shares,permalink_url,from,limit{$limit}&access_token=$token");
   }
 }