+++ CHANGELOG.md
@@ -1,3 +1,11 @@
-## [0.0.1] - TODO: Add release date.
+## [0.0.2] - 2021.01.14
 
-* TODO: Describe initial release.
+* add example
+* add Doc
+* add custom limit
+* add custom waiting widget
+* add custom noDataOrErrow widget
+
+## [0.0.1] - 2021.01.14
+
+* inital Release